backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
psycopg2-binary
openai
openai-agents
python-dotenv
bcrypt==4.2.1
PyJWT==2.9.0
